Let animations fully arrive at target position
Previously we got very close but not quite, due to the way zooming
affects the values needed for panning. So you would hit a key and
then hitting it again would cause another slight adjustment.
This was also the reason the zero key wasn't quite getting fully
back to the zero view when zooming was involved.